Add files via upload
plant page with upvoting and downvoting, also added two variables in the plantProDB for them. However I don't know how to make changes of DB once I clicked on the button in html. I am thinking of users can click on the button Like and the count for upvoting in DB will add 1.